CHORE:refactor - relocate cache gettr from api.py to cache.py
