+++ app/main/views.py
@@ -3,7 +3,7 @@ import logging
 from django.http import HttpResponse
 from django.template import loader
 
-from app.main.api import (
+from app.main.cache import (
     fetch_global_notifications,
     get_explore_the_collection,
 )
